+++ pyfr/plugins/soln/dtstats.py
@@ -1,5 +1,6 @@
 from pyfr.mpiutil import get_comm_rank_root
-from pyfr.plugins.base import BaseSolnPlugin, init_csv
+from pyfr.plugins.common import init_csv
+from pyfr.plugins.soln.base import BaseSolnPlugin
 
 
 class DtStatsPlugin(BaseSolnPlugin):
