+++ pyfr/plugins/soln/fluidforce.py
@@ -2,8 +2,9 @@ import numpy as np
 
 from pyfr.cache import memoize
 from pyfr.mpiutil import get_comm_rank_root, mpi
-from pyfr.plugins.base import (BackendMixin, BaseSolnPlugin, DatasetAppender,
-                               PublishMixin, init_csv, open_hdf5_a)
+from pyfr.plugins.common import DatasetAppender, init_csv, open_hdf5_a
+from pyfr.plugins.mixins import BackendMixin, PublishMixin
+from pyfr.plugins.soln.base import BaseSolnPlugin
 from pyfr.quadrules.surface import SurfaceIntegrator
