+++ pyfr/plugins/soln/writer.py
@@ -1,6 +1,7 @@
 from pyfr.inifile import Inifile
 from pyfr.mpiutil import get_comm_rank_root
-from pyfr.plugins.base import BaseSolnPlugin, PostactionMixin, RegionMixin
+from pyfr.plugins.mixins import PostactionMixin, RegionMixin
+from pyfr.plugins.soln.base import BaseSolnPlugin
 from pyfr.writers.native import NativeWriter
 from pyfr.util import first
